\name{IC2-package}
\alias{IC2-package}
\alias{IC2}
\docType{package}
\title{
Inequality and Concentration Index and Curves
}
\description{
The package \pkg{IC2} implements the computation of some indices of inequality
and concentration. For each index, it provides decomposition between subgroups.
Plotting of Lorenz and concentration curves are also available.
}
\details{
Three family of inequality indices are available: extended Gini, Atkinson and
Generalized Entropy. Except for GEI, two different forms of decomposition are
available. Ordinary as well as generalized Lorenz curves can be drawn. Sampling
weights can be used.
}
\author{Didier Plat}
\keyword{univar}
\keyword{multivariate}